# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "api_key",
        "type": "apiKey"
      },
      "APIKeyHeader": {
        "in": "header",
        "name": "api_key",
        "type": "apiKey"
      },
      "APIKeyQuery": {
        "in": "query",
        "name": "api_key",
        "type": "apiKey"
      },
      "ApplicationCredential": {
        "description": "Application credential issued by the operator with `custom-domain credential issue`. Send it as `Authorization: Bearer cd_...`. The application is derived from the credential; it is never taken from the request.",
        "scheme": "bearer",
        "type": "http"
      },
      "OperatorToken": {
        "description": "The deployment's OPERATOR_API_TOKEN, as `Authorization: Bearer <token>`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\nCustom domains for multi-tenant SaaS. Applications register customer hostnames\nfor their workspaces, hand the returned DNS records to the customer, and are\ntold when the hostname is verified, certified and serving.\n\nApplications are registered by the operator (`custom-domain application create`)\nand receive credentials from the operator (`custom-domain credential issue`).\nThere is no self-service registration endpoint in v1. Every request is scoped\nto the application that owns the presented credential.\n\nThe endpoints under `/domains` without a version prefix are the legacy\nsingle-application API. They are deprecated and disabled by setting\n`ENABLE_LEGACY_API=false`.\n",
    "title": "Custom Domain API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/operator/v1/applications": {
      "get": {
        "operationId": "list_applications_operator_v1_applications_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApplicationResource"
                  },
                  "title": "Response List Applications Operator V1 Applications Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or wrong operator token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The client address may not use it."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found, or the operator API is off."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request is not valid."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too many failed tokens from this client."
          }
        },
        "security": [
          {
            "OperatorToken": []
          }
        ],
        "summary": "List Applications",
        "tags": [
          "Operator"
        ]
      },
      "post": {
        "operationId": "create_application_operator_v1_applications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or wrong operator token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The client address may not use it."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found, or the operator API is off."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request is not valid."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too many failed tokens from this client."
          }
        },
        "security": [
          {
            "OperatorToken": []
          }
        ],
        "summary": "Create Application",
        "tags": [
          "Operator"
        ]
      }
    },
...
    "/operator/v1/applications/{slug}/assertion-keys": {
      "get": {
        "description": "The application's own assertion keys (no secrets), and which one signs now.",
        "operationId": "list_assertion_keys_operator_v1_applications__slug__assertion_keys_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssertionKeys"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or wrong operator token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The client address may not use it."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found, or the operator API is off."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request is not valid."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too many failed tokens from this client."
          }
        },
        "security": [
          {
            "OperatorToken": []
          }
        ],
        "summary": "List Assertion Keys",
        "tags": [
          "Operator"
        ]
      },
      "post": {
        "description": "Issue the application's next assertion key; the secret is in this response only.",
        "operationId": "issue_assertion_key_operator_v1_applications__slug__assertion_keys_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AssertionKeyCreate"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewAssertionKey"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or wrong operator token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The client address may not use it."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found, or the operator API is off."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request is not valid."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too many failed tokens from this client."
          }
        },
        "security": [
          {
            "OperatorToken": []
          }
        ],
        "summary": "Issue Assertion Key",
        "tags": [
          "Operator"
        ]
      }
    },
...
    "/operator/v1/applications/{slug}/credentials/{credential_id}": {
      "delete": {
        "operationId": "delete_credential_operator_v1_applications__slug__credentials__credential_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "credential_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Credential Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or wrong operator token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The client address may not use it."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found, or the operator API is off."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request is not valid."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too many failed tokens from this client."
          }
        },
        "security": [
          {
            "OperatorToken": []
          }
        ],
        "summary": "Delete Credential",
        "tags": [
          "Operator"
        ]
      }
    },
...
  },
...
}